`vim` にパスが通っていた場合、その vim を使用するように修正。
現状、 `--nvim` オプションをつけても `vim` が起動する。

+++ devcontainer/devcontainerStartVimArgs_darwin_arm64.go
@@ -2,17 +2,26 @@
 
 package devcontainer
 
-func devcontainerStartVimArgs(containerID string, workspaceFolder string, vimFileName string) []string {
-	return []string{
-		"exec",
-		"--container-id",
-		containerID,
-		"--workspace-folder",
-		workspaceFolder,
-		"sh",
-		"-c",
-		"cd /; tar zxf ./" + vimFileName + " -C ~/ > /dev/null; cd ~; sudo rm -rf ~/vim-static; mv $(ls -d ~/vim-*-aarch64) ~/vim-static;~/vim-static/AppRun --cmd \"let g:devcontainer_vim = v:true\" -S /SendToTcp.vim -S /vimrc"}
-
+func devcontainerStartVimArgs(containerID string, workspaceFolder string, vimFileName string, useSystemVim bool) []string {
+	if useSystemVim {
+		return []string{
+			"exec",
+			"--container-id",
+			containerID,
+			"--workspace-folder",
+			workspaceFolder,
+			"sh",
+			"-c",
+			"vim --cmd \"let g:devcontainer_vim = v:true\" -S /SendToTcp.vim -S /vimrc"}
+	} else {
+		return []string{
+			"exec",
+			"--container-id",
+			containerID,
+			"--workspace-folder",
+			workspaceFolder,
+			"sh",
+			"-c",
+			"cd /; tar zxf ./" + vimFileName + " -C ~/ > /dev/null; cd ~; sudo rm -rf ~/vim-static; mv $(ls -d ~/vim-*-aarch64) ~/vim-static;~/vim-static/AppRun --cmd \"let g:devcontainer_vim = v:true\" -S /SendToTcp.vim -S /vimrc"}
+	}
 }
-
-
